+++ Tests/LegacyUnit/FrontEnd/AbstractViewTest.php
@@ -28,6 +28,7 @@ final class AbstractViewTest extends TestCase
     {
         $this->testingFramework = new TestingFramework('tx_seminars');
         $rootPageUid = $this->testingFramework->createFrontEndPage();
+        $this->testingFramework->changeRecord('pages', $rootPageUid, ['slug' => '/home']);
         $this->testingFramework->createFakeFrontEnd($rootPageUid);
         $this->subject = new TestingView(
             ['templateFile' => 'EXT:seminars/Resources/Private/Templates/FrontEnd/FrontEnd.html'],